move runtime apis to lib rs to try fix metadata generation bug
