+++ runtime/src/lib.rs
@@ -85,7 +85,7 @@
 #![allow(non_local_definitions)]
 
 // The runtime is split into its components
-pub mod apis;
+//pub mod apis;
 pub mod configs;
 pub mod migrations;
 pub mod offchain;
@@ -93,9 +93,6 @@ pub mod version;
 
 /// Helpers to handle variant flags
 pub mod variants;
-
-pub use apis::RuntimeApi;
-use apis::_InternalImplRuntimeApis;
 
 pub use version::VERSION;
 
@@ -142,7 +139,13 @@ use frame_support::{
 };
 use pallet_session::historical as pallet_session_historical;
 
+use frame_support::traits::KeyOwnerProofSystem;
+use pallet_staking::migrations::v17::MigrateDisabledToSession;
+use scale_codec::Encode;
 use sp_runtime::generic;
+use sp_runtime::traits::Block as BlockT;
+use sp_runtime::transaction_validity::{TransactionSource, TransactionValidity};
+use sp_runtime::KeyTypeId;
 use sp_std::prelude::*;
 
 pub use time_primitives::{
@@ -613,7 +616,8 @@ mod runtime {
 
 // All migrations executed on runtime upgrade implementing `OnRuntimeUpgrade`.
 type Migrations = (
-	pallet_session::migrations::v1::MigrateV0ToV1<Runtime, Historical>,
+	// Use the MigrateDisabledToSession implementation which implements MigrateDisabledValidators
+	pallet_session::migrations::v1::MigrateV0ToV1<Runtime, MigrateDisabledToSession<Runtime>>,
 	migrations::staking::StakingCurrencyMigration<Runtime>,
 );
 
@@ -644,5 +648,428 @@ mod core_tests {
 			 If the limit is too strong, maybe consider increase the limit.",
 			size,
 		);
+	}
+}
+
+sp_api::impl_runtime_apis! {
+
+	impl frame_system_rpc_runtime_api::AccountNonceApi<Block, AccountId, Nonce> for Runtime {
+		fn account_nonce(account: AccountId) -> Nonce {
+			System::account_nonce(account)
+		}
+	}
+
+	impl sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block> for Runtime {
+		fn validate_transaction(
+			source: TransactionSource,
+			tx: <Block as BlockT>::Extrinsic,
+			block_hash: <Block as BlockT>::Hash,
+		) -> TransactionValidity {
+			Executive::validate_transaction(source, tx, block_hash)
+		}
+	}
+
+	impl sp_api::Core<Block> for Runtime {
+		fn version() -> sp_version::RuntimeVersion {
+			VERSION
+		}
+
+		fn execute_block(block: Block) {
+			Executive::execute_block(block);
+		}
+
+		fn initialize_block(header: &<Block as sp_runtime::traits::Block>::Header) -> sp_runtime::ExtrinsicInclusionMode {
+			Executive::initialize_block(header)
+		}
+	}
+
+	impl sp_api::Metadata<Block> for Runtime {
+		fn metadata() -> sp_core::OpaqueMetadata {
+			sp_core::OpaqueMetadata::new(Runtime::metadata().into())
+		}
+
+		fn metadata_at_version(version: u32) -> Option<sp_core::OpaqueMetadata> {
+			Runtime::metadata_at_version(version)
+		}
+
+		fn metadata_versions() -> sp_std::vec::Vec<u32> {
+			Runtime::metadata_versions()
+		}
+	}
+
+	impl sp_authority_discovery::AuthorityDiscoveryApi<Block> for Runtime {
+		fn authorities() -> Vec<sp_authority_discovery::AuthorityId> {
+			AuthorityDiscovery::authorities()
+		}
+	}
+
+	impl sp_block_builder::BlockBuilder<Block> for Runtime {
+		fn apply_extrinsic(extrinsic: <Block as sp_runtime::traits::Block>::Extrinsic) -> sp_runtime::ApplyExtrinsicResult {
+			Executive::apply_extrinsic(extrinsic)
+		}
+
+		fn finalize_block() -> <Block as sp_runtime::traits::Block>::Header {
+			Executive::finalize_block()
+		}
+
+		fn inherent_extrinsics(data: sp_inherents::InherentData) -> Vec<<Block as sp_runtime::traits::Block>::Extrinsic> {
+			data.create_extrinsics()
+		}
+
+		fn check_inherents(block: Block, data: sp_inherents::InherentData) -> sp_inherents::CheckInherentsResult {
+			data.check_extrinsics(&block)
+		}
+	}
+
+	impl sp_consensus_babe::BabeApi<Block> for Runtime {
+		fn configuration() -> sp_consensus_babe::BabeConfiguration {
+			let epoch_config = Babe::epoch_config().unwrap_or(BABE_GENESIS_EPOCH_CONFIG);
+			sp_consensus_babe::BabeConfiguration {
+				slot_duration: Babe::slot_duration(),
+				epoch_length: EpochDuration::get(),
+				c: epoch_config.c,
+				authorities: Babe::authorities().to_vec(),
+				randomness: Babe::randomness(),
+				allowed_slots: epoch_config.allowed_slots,
+			}
+		}
+
+		fn current_epoch_start() -> sp_consensus_babe::Slot {
+			Babe::current_epoch_start()
+		}
+
+		fn current_epoch() -> sp_consensus_babe::Epoch {
+			Babe::current_epoch()
+		}
+
+		fn next_epoch() -> sp_consensus_babe::Epoch {
+			Babe::next_epoch()
+		}
+
+		fn generate_key_ownership_proof(
+			_slot: sp_consensus_babe::Slot,
+			authority_id: sp_consensus_babe::AuthorityId,
+		) -> Option<sp_consensus_babe::OpaqueKeyOwnershipProof> {
+			Historical::prove((sp_consensus_babe::KEY_TYPE, authority_id))
+				.map(|p| p.encode())
+				.map(sp_consensus_babe::OpaqueKeyOwnershipProof::new)
+		}
+
+		fn submit_report_equivocation_unsigned_extrinsic(
+			equivocation_proof: sp_consensus_babe::EquivocationProof<<Block as sp_runtime::traits::Block>::Header>,
+			key_owner_proof: sp_consensus_babe::OpaqueKeyOwnershipProof,
+		) -> Option<()> {
+			Babe::submit_unsigned_equivocation_report(
+				equivocation_proof,
+				key_owner_proof.decode()?,
+			)
+		}
+	}
+
+	impl sp_consensus_grandpa::GrandpaApi<Block> for Runtime {
+		fn grandpa_authorities() -> sp_consensus_grandpa::AuthorityList {
+			Grandpa::grandpa_authorities()
+		}
+
+		fn current_set_id() -> sp_consensus_grandpa::SetId {
+			Grandpa::current_set_id()
+		}
+
+		fn submit_report_equivocation_unsigned_extrinsic(
+			equivocation_proof: sp_consensus_grandpa::EquivocationProof<
+				<Block as sp_runtime::traits::Block>::Hash,
+				sp_runtime::traits::NumberFor<Block>,
+			>,
+			key_owner_proof: sp_consensus_grandpa::OpaqueKeyOwnershipProof,
+		) -> Option<()> {
+			let key_owner_proof = key_owner_proof.decode()?;
+
+			Grandpa::submit_unsigned_equivocation_report(
+				equivocation_proof,
+				key_owner_proof,
+			)
+		}
+
+		fn generate_key_ownership_proof(
+			_set_id: sp_consensus_grandpa::SetId,
+			authority_id: sp_consensus_grandpa::AuthorityId,
+		) -> Option<sp_consensus_grandpa::OpaqueKeyOwnershipProof> {
+			Historical::prove((sp_consensus_grandpa::KEY_TYPE, authority_id))
+				.map(|p| p.encode())
+				.map(sp_consensus_grandpa::OpaqueKeyOwnershipProof::new)
+		}
+	}
+
+	impl sp_offchain::OffchainWorkerApi<Block> for Runtime {
+		fn offchain_worker(header: &<Block as sp_runtime::traits::Block>::Header) {
+			Executive::offchain_worker(header)
+		}
+	}
+
+	impl sp_session::SessionKeys<Block> for Runtime {
+		fn generate_session_keys(seed: Option<Vec<u8>>) -> Vec<u8> {
+			SessionKeys::generate(seed)
+		}
+
+		fn decode_session_keys(encoded: Vec<u8>) -> Option<Vec<(Vec<u8>, KeyTypeId)>> {
+			SessionKeys::decode_into_raw_public_keys(&encoded)
+		}
+	}
+
+	impl pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance> for Runtime {
+		fn query_info(uxt: <Block as sp_runtime::traits::Block>::Extrinsic, len: u32) -> pallet_transaction_payment::RuntimeDispatchInfo<Balance> {
+			TransactionPayment::query_info(uxt, len)
+		}
+
+		fn query_fee_details(uxt: <Block as sp_runtime::traits::Block>::Extrinsic, len: u32) -> pallet_transaction_payment::FeeDetails<Balance> {
+			TransactionPayment::query_fee_details(uxt, len)
+		}
+
+		fn query_weight_to_fee(weight: Weight) -> Balance {
+			TransactionPayment::weight_to_fee(weight)
+		}
+
+		fn query_length_to_fee(length: u32) -> Balance {
+			TransactionPayment::length_to_fee(length)
+		}
+	}
+
+	impl pallet_nomination_pools_runtime_api::NominationPoolsApi<Block, AccountId, Balance> for Runtime {
+		fn pending_rewards(member: AccountId) -> Balance {
+			NominationPools::api_pending_rewards(member).unwrap_or_default()
+		}
+
+		fn points_to_balance(pool_id: u32, points: Balance) -> Balance {
+			NominationPools::api_points_to_balance(pool_id, points)
+		}
+
+		fn balance_to_points(pool_id: u32, balance: Balance) -> Balance {
+			NominationPools::api_balance_to_points(pool_id, balance)
+		}
+
+		fn pool_pending_slash(pool_id: u32) -> Balance {
+			NominationPools::api_pool_pending_slash(pool_id)
+		}
+
+		fn member_pending_slash(member: AccountId) -> Balance {
+			NominationPools::api_member_pending_slash(member)
+		}
+
+		fn pool_needs_delegate_migration(pool_id: u32) -> bool {
+			NominationPools::api_pool_needs_delegate_migration(pool_id)
+		}
+
+		fn member_needs_delegate_migration(member: AccountId) -> bool {
+			NominationPools::api_member_needs_delegate_migration(member)
+		}
+
+		fn member_total_balance(member: AccountId) -> Balance {
+			NominationPools::api_member_total_balance(member)
+		}
+
+		fn pool_balance(pool_id: u32) -> Balance {
+			NominationPools::api_pool_balance(pool_id)
+		}
+
+		fn pool_accounts(pool_id: u32) -> (AccountId, AccountId) {
+			NominationPools::api_pool_accounts(pool_id)
+		}
+	}
+
+	impl pallet_staking_runtime_api::StakingApi<Block, Balance, AccountId> for Runtime {
+		fn nominations_quota(balance: Balance) -> u32 {
+			Staking::api_nominations_quota(balance)
+		}
+
+		fn eras_stakers_page_count(era: u32, account: AccountId) -> u32 {
+			Staking::api_eras_stakers_page_count(era, account)
+		}
+
+		fn pending_rewards(era: u32, account: AccountId) -> bool {
+			Staking::api_pending_rewards(era, account)
+		}
+	}
+
+	#[cfg(feature = "testnet")]
+	impl time_primitives::NetworksApi<Block> for Runtime {
+		fn get_network(network_id: NetworkId) -> Option<time_primitives::ChainName> {
+			Networks::get_network(network_id)
+		}
+
+		fn get_gateway(network: NetworkId) -> Option<time_primitives::Address32> {
+			Networks::gateway(network)
+		}
+
+		fn get_cctp_contracts(network: NetworkId) -> Option<time_primitives::CctpContracts> {
+			Networks::get_cctp_contracts(network)
+		}
+
+		fn get_cctp_url(network: NetworkId) -> Option<time_primitives::CctpUrl> {
+			Networks::get_cctp_url(network)
+		}
+	}
+
+	#[cfg(feature = "testnet")]
+	impl time_primitives::ShardsApi<Block> for Runtime {
+		fn get_shards(account: &AccountId) -> Vec<time_primitives::ShardId> {
+			Shards::get_shards(account)
+		}
+
+		fn get_shard_members(shard_id: time_primitives::ShardId) -> Vec<(AccountId, time_primitives::MemberStatus)> {
+			Shards::get_shard_members(shard_id)
+		}
+
+		fn get_shard_threshold(shard_id: time_primitives::ShardId) -> u16 {
+			Shards::get_shard_threshold(shard_id)
+		}
+
+		fn get_shard_status(shard_id: time_primitives::ShardId) -> time_primitives::ShardStatus {
+			Shards::get_shard_status(shard_id)
+		}
+
+		fn get_shard_commitment(shard_id: time_primitives::ShardId) -> Option<time_primitives::Commitment> {
+			Shards::get_shard_commitment(shard_id)
+		}
+	}
+
+	#[cfg(feature = "testnet")]
+	impl time_primitives::TasksApi<Block> for Runtime {
+		fn get_shard_tasks(shard_id: time_primitives::ShardId) -> Vec<time_primitives::TaskId> {
+			Tasks::get_shard_tasks(shard_id)
+		}
+
+		fn get_task(task_id: time_primitives::TaskId) -> Option<time_primitives::Task>{
+			Tasks::get_task(task_id)
+		}
+
+		fn get_task_result(task_id: time_primitives::TaskId) -> Option<Result<(), time_primitives::ErrorMsg>>{
+			Tasks::get_task_result(task_id)
+		}
+
+		fn get_task_shard(task_id: time_primitives::TaskId) -> Option<time_primitives::ShardId>{
+			Tasks::get_task_shard(task_id)
+		}
+
+		fn get_batch_message(batch_id: time_primitives::BatchId) -> Option<time_primitives::GatewayMessage> {
+			Tasks::get_batch_message(batch_id)
+		}
+
+		fn get_failed_tasks() -> Vec<time_primitives::TaskId> {
+			Tasks::get_failed_tasks()
+		}
+	}
+
+	#[cfg(feature = "testnet")]
+	impl time_primitives::SubmitTransactionApi<Block> for Runtime {
+		fn submit_transaction(encoded_transaction: Vec<u8>) -> Result<(), ()> {
+			sp_io::offchain::submit_transaction(encoded_transaction)
+		}
+	}
+
+	/// - __genesis-builder__: support generation of custom genesis
+	#[cfg(feature = "genesis-builder")]
+	impl sp_genesis_builder::GenesisBuilder<Block> for Runtime {
+		fn build_state(config: Vec<u8>) -> sp_genesis_builder::Result {
+			use frame_support::genesis_builder_helper::build_state;
+			build_state::<RuntimeGenesisConfig>(config)
+		}
+
+		fn get_preset(id: &Option<sp_genesis_builder::PresetId>) -> Option<Vec<u8>> {
+			use frame_support::genesis_builder_helper::get_preset;
+			get_preset::<RuntimeGenesisConfig>(id, |_| None)
+		}
+
+		fn preset_names() -> Vec<sp_genesis_builder::PresetId> {
+			vec![]
+		}
+	}
+
+	/// - __runtime-benchmarks__: support runtime benchmarking
+	#[cfg(feature = "runtime-benchmarks")]
+	impl frame_benchmarking::Benchmark<Block> for Runtime {
+		fn benchmark_metadata(extra: bool) -> (
+			Vec<frame_benchmarking::BenchmarkList>,
+			Vec<frame_support::traits::StorageInfo>,
+		) {
+			use frame_benchmarking::{baseline, Benchmarking, BenchmarkList};
+			use frame_support::traits::StorageInfoTrait;
+
+			// Trying to add benchmarks directly to the Session Pallet caused cyclic dependency
+			// issues. To get around that, we separated the Session benchmarks into its own crate,
+			// which is why we need these two lines below.
+			use pallet_session_benchmarking::Pallet as SessionBench;
+			use pallet_nomination_pools_benchmarking::Pallet as NominationPoolsBench;
+			use pallet_offences_benchmarking::Pallet as OffencesBench;
+			use pallet_election_provider_support_benchmarking::Pallet as EPSBench;
+			use frame_system_benchmarking::Pallet as SystemBench;
+			use frame_system_benchmarking::extensions::Pallet as SystemExtensionsBench;
+			use baseline::Pallet as BaselineBench;
+
+			// Import substrate macros created by macros (and all pallets by effects)
+			use crate::*;
+
+			let mut list = Vec::<BenchmarkList>::new();
+			list_benchmarks!(list, extra);
+			let storage_info = AllPalletsWithSystem::storage_info();
+			(list, storage_info)
+		}
+
+		fn dispatch_benchmark(
+			config: frame_benchmarking::BenchmarkConfig
+		) -> Result<Vec<frame_benchmarking::BenchmarkBatch>, String> {
+			use frame_benchmarking::{baseline, Benchmarking, BenchmarkBatch};
+
+			// Trying to add benchmarks directly to the Session Pallet caused cyclic dependency
+			// issues. To get around that, we separated the Session benchmarks into its own crate,
+			// which is why we need these two lines below.
+			use pallet_session_benchmarking::Pallet as SessionBench;
+			use pallet_nomination_pools_benchmarking::Pallet as NominationPoolsBench;
+			use pallet_offences_benchmarking::Pallet as OffencesBench;
+			use pallet_election_provider_support_benchmarking::Pallet as EPSBench;
+			use frame_system_benchmarking::Pallet as SystemBench;
+			use frame_system_benchmarking::extensions::Pallet as SystemExtensionsBench;
+			use baseline::Pallet as BaselineBench;
+
+			impl pallet_session_benchmarking::Config for Runtime {}
+			impl pallet_nomination_pools_benchmarking::Config for Runtime {}
+			impl pallet_offences_benchmarking::Config for Runtime {}
+			impl pallet_election_provider_support_benchmarking::Config for Runtime {}
+			impl frame_system_benchmarking::Config for Runtime {}
+			impl baseline::Config for Runtime {}
+
+			use frame_support::traits::{TrackedStorageKey, WhitelistedStorageKeys};
+			let whitelist: Vec<TrackedStorageKey> = AllPalletsWithSystem::whitelisted_storage_keys();
+
+			// Import substrate macros created by macros (and all pallets by effects)
+			use crate::*;
+
+			let mut batches = Vec::<BenchmarkBatch>::new();
+			let params = (&config, &whitelist);
+			add_benchmarks!(params, batches);
+			Ok(batches)
+		}
+	}
+
+	/// - __try-runtime__: support try runtime testing
+	#[cfg(feature = "try-runtime")]
+	impl frame_try_runtime::TryRuntime<Block> for Runtime {
+		fn on_runtime_upgrade(checks: frame_try_runtime::UpgradeCheckSelect) -> (Weight, Weight) {
+			// NOTE: intentional unwrap: we don't want to propagate the error backwards, and want to
+			// have a backtrace here. If any of the pre/post migration checks fail, we shall stop
+			// right here and right now.
+			let weight = Executive::try_runtime_upgrade(checks).unwrap();
+			(weight, crate::RuntimeBlockWeights::get().max_block)
+		}
+
+		fn execute_block(
+			block: Block,
+			state_root_check: bool,
+			signature_check: bool,
+			select: frame_try_runtime::TryStateSelect
+		) -> Weight {
+			// NOTE: intentional unwrap: we don't want to propagate the error backwards, and want to
+			// have a backtrace here.
+			Executive::try_execute_block(block, state_root_check, signature_check, select).unwrap()
+		}
 	}
 }
